Menachem's Initial commit
This commit includes the TODO data structure implementation and the system calls add_TODO and read_TODO.
TODO_copy copies all the files to their target paths in the custom kernel.
make_kernel is a script that goes through all the kernel compilation stages.
short_make_kernel is a script that skips some of the stages (can be done if no *.h or *.S file was changed)